requirements.txt
anthropic>=0.25.0
moviepy>=2.1.2
Pillow>=10.0.0
requests>=2.31.0
gtts>=2.4.0
numpy>=1.24.0
